Rename Operations => Actions
Actions is a more fitting name for the route this API is taking. Operations came from the `Savon::Client#operations` method which provides a list of possible SOAP operations you can call.

I prefer to think of this as an "action" rather than an "operation" within the confines of this gem.